jsp 17일차: 로그인 로그아웃 회원관리 예제

+++ memberPro/WebContent/member_list.jsp
@@ -2,6 +2,7 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8"
     pageEncoding="UTF-8"%>
 <%@ page import="java.util.*" %>
+<%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core"%>
     
 <!DOCTYPE html>
 <html>
@@ -31,15 +32,9 @@
 		</tr>
 		<c:forEach var="member" items="${memberList}">
 		<tr>
-			<td>
-				<a href="memberViewAction.do?id=${member.member_id}">
-					${mamber.member_id}
-				</a>
-			</td>
-			<td>
-			<a href="memberDeleteAction.do?id=${member.member_id}">삭제</a>
-			</td>
-			</tr>
+			<td><a href="memberViewAction.do?id=${member.member_id}">${member.member_id}</a></td>
+			<td><a href="memberDeleteAction.do?id=${member.member_id}"onclick="return confirm('삭제하시겠습니까?')">삭제</a></td>
+		</tr>
 		</c:forEach>	
 	</table>
 </section>
